+++ toml/__init__.py
@@ -1,4 +1,4 @@
 __author__ = "elpekenin"
-__version__ = (0, 1, 8)
+__version__ = (0, 1, 9)
 
 from ._toml import *