+++ toml/_toml.py
@@ -13,204 +13,243 @@ class TOMLError(Exception):
 
 
 class Tokens:
+    """Different strings with "special" meaning."""
+
     OPENING_BRACKET = "["
     CLOSING_BRACKET = "]"
-    SINGLE_QUOTE = "'"
-    DOUBLE_QUOTE = '"'
+    QUOTE = "'"
+    DQUOTE = '"'
+    TRIPLE_QUOTE = "'''"
+    TRIPLE_DQUOTE = '"""'
     EQUAL_SIGN = "="
     COMMENT = "#"
     COMMA = ","
+    BACKSLASH = "\\"
 
-    QUOTES = {
-        SINGLE_QUOTE,
-        DOUBLE_QUOTE,
-    }
+    # triple has to be first, as regular quotes would match too
+    # thus, this also needs to be a list, to maintain order
+    QUOTES = [
+        TRIPLE_QUOTE,
+        TRIPLE_DQUOTE,
+        QUOTE,
+        DQUOTE,
+    ]
 
     ALL = {
         OPENING_BRACKET,
         CLOSING_BRACKET,
-        SINGLE_QUOTE,
-        DOUBLE_QUOTE,
+        QUOTE,
+        DQUOTE,
+        TRIPLE_QUOTE,
+        TRIPLE_DQUOTE,
         EQUAL_SIGN,
+        COMMENT,
         COMMA,
+        BACKSLASH,
     }
 
 
-class LineInfo:
+class ParsedLine:
     """Cleanup raw line's content and find tokens on it."""
 
     line: str
     """Clean line (strip()'ed and comments removed)."""
 
-    tokens: dict
+    tokens: dict[str, list[int]]
     """Mapping from tokens to the position(s) where they are found on the line."""
 
-    assignment: int = -1
-    """Position of the equal sign, -1 if none."""
+    def __str__(self) -> str:
+        return f"line={self.line!r}, tokens={self.tokens!r}"
 
-    had_comment: bool = False
-    """Whether the line contained a comment."""
+    __repr__ = __str__
 
     def __init__(self, __line: str):
         self.line = ""
         self.tokens = {t: [] for t in Tokens.ALL}
 
-        in_quotes = False
         quote_token = None
 
-        # TODO: Rework into while, for better string parsing (quote/escapes)
-        for i, char in enumerate(__line.lstrip()):
-            # upon finding a comment (not in a quoted string), quit
-            if char == Tokens.COMMENT and not in_quotes:
+        stripped = __line.strip()
+        length = len(stripped)
+
+        i = 0
+        while i < length:
+            char = stripped[i]
+
+            # upon finding a comment (not in a string), quit
+            if (
+                char == Tokens.COMMENT
+                and quote_token is None
+            ):
                 # clean trailing spaces
                 self.line = self.line.rstrip()
-                self.had_comment = True
                 return
 
-            # add current char to "clean" string
-            self.line += char
+            token, string, offset = Parser.string(stripped[i:])
+            if string:
+                self.tokens[token].append(i)
+                self.tokens[token].append(i + offset)
 
-            if char in Tokens.QUOTES:
-                # opening quote
-                if not quote_token:
-                    in_quotes = True
-                    quote_token = char
-                    self.tokens[char].append(i)
-                    continue
+                i += offset
+                self.line += string
 
-                # closing quote
-                elif char == quote_token:
-                    in_quotes = False
-                    quote_token = None
-                    self.tokens[char].append(i)
-                    continue
-
-            # no token data to be stored if we are in a string
-            if in_quotes:
                 continue
 
-            # metadata about tokens
-            if char in Tokens.ALL:
-                self.tokens[char].append(i)
+            i += 1
+            self.line += char
 
-                # assignment location
-                if char == Tokens.EQUAL_SIGN and self.assignment == -1:
-                    self.assignment = i
+            # store tokens' positions
+            if char in Tokens.ALL:
+                self.tokens[char].append(len(self.line) - 1)
 
         # clean trailing spaces
         self.line = self.line.rstrip()
 
+    def is_empty(self) -> bool:
+        """Whether this line contains anything."""
+        return not bool(self.line)
 
-class SyntaxChecker:
-    """Some basic syntax rules based on the tokens found."""
+    def key_value(self) -> tuple[str, str]:
+        """Get the key and value on this line (ie: split on equal sign)."""
 
-    @staticmethod
-    def check(info: LineInfo) -> Optional[str]:
-        """Run some checks."""
+        assert len(self.tokens[Tokens.EQUAL_SIGN]), "How did we end up on key_value with len(EQUAL) != 1"
 
-        #################
-        # String checks #
-        #################
+        split_at = self.tokens[Tokens.EQUAL_SIGN][0]
+        key = self.line[:split_at].strip()
+        value = self.line[split_at + 1 :].strip()
 
-        single = len(info.tokens[Tokens.SINGLE_QUOTE])
-        double = len(info.tokens[Tokens.DOUBLE_QUOTE])
-
-        # every openning quote should be closed too
-        if single % 2 != 0 or double % 2 != 0:
-            return "Malformed string, check out your quotes"
-
-        #######################
-        # Table or assignment #
-        #######################
-        is_assignment = info.assignment != -1
-        is_table_setter = (
-            info.line[0] == Tokens.OPENING_BRACKET
-            and info.line[-1] == Tokens.CLOSING_BRACKET
-        )
-
-        if not (is_assignment or is_table_setter):
-            return "Line has to contain either an assignment or table setter"
-
-        if is_assignment and is_table_setter:
-            return "Line cant be an assignment and table setter at the same time"
-
-        ##############
-        # Assignment #
-        ##############
-        if is_assignment and not len(info.line) > (info.assignment + 1):
-            return "Invalid assignment, nothing after equal sign"
-
-        # If we got here, everything was correct
-        # Empty string => No exception raised
-        return ""
+        return key, value
 
 
 class Parser:
-    """Extract information from a TOML."""
+    """Get Python values out of strings."""
 
-    data: Dotty
-    """Dotty dict where the information from the TOML is stored."""
+    _SEPARATOR = "__"
+    """Used on Dotty to mark nesting."""
 
-    _table: str
-    """Current table ([table], [another.table]) of the parser."""
-
-    def __init__(self, __text: str):
-        """Parse incoming TOML."""
-
-        self.data = Dotty()
-        self._table = ""
-
-        lines = __text.replace("\r", "").split("\n")
-        for i, line in enumerate(lines, 1):
-            # TODO: Remove comments before processing the line
-            self._parse_line(i, line)
-
-    def _is_quoted(self, __val: str) -> bool:
-        """Check if a string is quoted."""
-        return __val[0] == __val[-1] and __val[0] in Tokens.QUOTES
-
-    def _sanitize_key(self, __key: str) -> str:
-        """Sanitize keys with quotes"""
-
-        parts = __key.split(".")
-        for i, raw_part in enumerate(parts):
-            # handle whitespace
-            clean_part = raw_part.strip()
-
-            # and quotes
-            if self._is_quoted(clean_part):
-                clean_part = clean_part[1:-1]
-
-            parts[i] = clean_part
-
-        final_key = ".".join(parts)
-        if self._is_quoted(final_key):
-            final_key = final_key[1:-1]
-        return final_key
-
-    def _parse_value(self, __value: str, __line_info: Optional[LineInfo] = None) -> Any:
+    @classmethod
+    def string(cls, __value: str) -> tuple[str, str, int]:
         """
-        (Try) Convert a string into a value.
+        Find the next **quoted** string in the input,
+        return it and how much the cursor has been moved.
+
+        Eg:
+        >>> string("'''hello'world'''")
+        >>> Tokens.TRIPLE_QUOTE, "hello'world", 17
+        """
+
+        quote_token = None
+        string = ""
+        i = 0
+
+        for token in Tokens.QUOTES:
+            sliced = __value[i : i+len(token)]
+            if sliced == token:
+                # opening quote
+                if quote_token is None:
+                    # store the string delimiter
+                    quote_token = token
+
+                    # the "clean" line should store single
+                    # quotes, not triple, thus append just current char
+                    string += token[0]
+
+                    # store this token
+                    i += len(token)
+
+                    break
+
+        # quote token not found, just exit
+        else:
+            return quote_token, string, i
+
+        # TODO: Handle escape sequences
+        length = len(__value)
+        while i < length:
+            char = __value[i]
+
+            # TODO: Better handling of this
+            if char == Tokens.BACKSLASH:
+                # blindly add the next char
+                string += __value[i + 1]
+
+                # and jump over it
+                i += 2
+                continue
+
+            # closing quote
+            sliced = __value[i : i+len(quote_token)]
+            if sliced == quote_token:
+                string += quote_token[0]
+
+                i += len(quote_token)
+
+                return quote_token, string, i
+
+            i += 1
+            string += char
+
+        # if we get down here, check that we did not had not found an opening
+        if quote_token is not None:
+            raise TOMLError("String was open but not closed.")
+
+    @classmethod
+    def key(cls, __key: str) -> str:
+        """
+        Sanitize keys with quotes, giving a Dotty-ready key.
+        """
+
+        # Note: The "__" here is Parser._SEPARATOR
+        #         input | output
+        #         ------|-------
+        #       foo.bar | foo__bar
+        #     "foo.bar" | foo.bar
+        # "foo.bar.baz" | foo.bar.baz
+        # "foo.bar".baz | foo.bar__baz  (is this possible)
+
+        parts = [None]
+        length = len(__key)
+
+        i = 0
+        while i < length:
+            _, string, offset = Parser.string(__key[i:])
+            if string:
+                i += offset
+                # NOTE: string is quoted, eg from a """hello"world"""
+                #       we get "hello\"world", stip head and trail quotes
+                #       but dont use .replace() as we might remove actual info
+                parts.append(string[1:-1])
+                continue
+
+            char = __key[i]
+            if char == ".":
+                parts.append(None)
+            else:
+                # do not add whitespace chars
+                if not char.isspace():
+                    # if last part is empty (None), replace it
+                    if parts[-1] is None:
+                        parts[-1] = ""
+
+                    parts[-1] += char
+
+            i += 1
+
+        # remove the (potential) empty strings that got added
+        return cls._SEPARATOR.join(part for part in parts if part is not None)
+
+    @classmethod
+    def value(cls, __value: str, __line_info: Optional[ParsedLine] = None) -> Any:
+        """
+        (Try) Convert a string into a Python value.
 
         Note: __line_info is only used when parsing lists
         """
 
         # quoted string, has to be first, to prevent casting it
-        if self._is_quoted(__value):
+        if Syntax.is_quoted(__value):
             # remove quotes
-            parsed_str = __value[1:-1]
-
-            # handle triple quotes (remove extra ones)
-            if parsed_str[0] == __value[0]:
-                if (
-                    parsed_str[0] == parsed_str[1]
-                    == parsed_str[-2] == parsed_str[-1]
-                ):
-                    return parsed_str[2:-2]
-
-                raise TOMLError("Malformed string")
-
-            return parsed_str
+            return __value[1:-1]
 
         # integer
         if __value.isdigit():
@@ -232,23 +271,20 @@ class Parser:
 
         # positive prefix, aka do nothing
         if __value[0] == "+":
-            return self._parse_value(__value[1:])
+            return cls.value(__value[1:])
 
         # negative numbers
         if __value[0] == "-":
             # spec does not allow negative numbers with base prefix
             if __value[1:3] not in ("0b", "0o", "0x"):
-                return -self._parse_value(__value[1:])
+                return -cls.value(__value[1:])
 
         # bool
         if __value in {"true", "false"}:
             return __value.lower() == "true"
 
         # array
-        if (
-            __value[0] == Tokens.OPENING_BRACKET
-            and __value[-1] == Tokens.CLOSING_BRACKET
-        ):
+        if Syntax.is_in_brackets(__value):
             if __line_info is None:
                 raise TOMLError("How did we end on array parsing without line info?")
 
@@ -258,7 +294,7 @@ class Parser:
             if len(opening) != len(closing):
                 raise TOMLError("Mismatched brackets.")
 
-            value, _ = self._parse_list(__line_info.line, opening[0] + 1)
+            value, _ = cls.list(__line_info.line, opening[0] + 1)
             return value
 
         # couldn't parse, raise Exception
@@ -266,15 +302,8 @@ class Parser:
             f"Couldn't parse value: `{__value}` (Hint, remember to wrap strings in quotes)"
         )
 
-    def _parse_assignment(self, __line_info: LineInfo) -> None:
-        split_at = __line_info.tokens[Tokens.EQUAL_SIGN][0]
-        _key = self._sanitize_key(__line_info.line[:split_at].strip())
-        _value = __line_info.line[split_at + 1 :].strip()
-
-        key = f"{self._table}.{_key}" if self._table else _key
-        self.data[key] = self._parse_value(_value, __line_info)
-
-    def _parse_list(self, __line: str, __start: int) -> tuple[list[Any], int]:
+    @classmethod
+    def list(cls, __line: str, __start: int) -> tuple[list[Any], int]:
         """
         Helper to parse a list.
         Returns parsed list + where next element starts
@@ -291,13 +320,13 @@ class Parser:
             if char == Tokens.CLOSING_BRACKET:
                 _text = text.strip()
                 if _text:
-                    elements.append(self._parse_value(_text))
+                    elements.append(cls.value(_text))
                 return elements, pos
 
             # parse list and update current position
             elif char == Tokens.OPENING_BRACKET:
                 text = ""
-                value, pos = self._parse_list(__line, pos)
+                value, pos = cls.list(__line, pos)
                 if value is not None:
                     elements.append(value)
 
@@ -305,7 +334,7 @@ class Parser:
             elif char == Tokens.COMMA:
                 _text = text.strip()
                 if _text:
-                    elements.append(self._parse_value(_text))
+                    elements.append(cls.value(_text))
                 text = ""
 
             # collect another char
@@ -315,31 +344,93 @@ class Parser:
         # how do we get here?
         return elements, pos
 
-    def _parse_line(self, __i: int, __line: str) -> None:
-        """Extract information from a line and add it to the Dotty."""
+    @classmethod
+    def toml(cls, __toml: str) -> Dotty:
+        """
+        Parse a whole TOML string.
+        """
 
-        # get information about this line
-        info = LineInfo(__line)
+        table = ""
+        data = Dotty(separator=cls._SEPARATOR)
 
-        # empty line => nothing to be done
-        if not info.line:
-            return
+        lines = __toml.replace("\r", "").split("\n")
+        for i, raw_line in enumerate(lines, 1):
+            parsed_line = ParsedLine(raw_line)
 
-        message = SyntaxChecker.check(info)
-        if message:
-            raise TOMLError(f"{message} in line {__i}")
+            # empty line => nothing to be done
+            if parsed_line.is_empty():
+                continue
 
-        # at this point, line should have content and correct syntax, this code can be rather dumb
-        # we can't strip or anything like that tho, indexes would be broken
+            message = Syntax.check(parsed_line)
+            if message:
+                raise TOMLError(f"{message} in line {i}")
 
-        # equal sign => assignment expresion
-        if Tokens.EQUAL_SIGN in info.line:
-            self._parse_assignment(info)
+            # at this point, line should have content and correct syntax, this code can be rather dumb
+            # we can't strip or anything like that tho, indexes would be broken
 
-        # no equal sign => table assignment, ie: [table]
-        else:
-            # remove "[" and "]", handle quotes/dots
-            self._table = self._sanitize_key(info.line[1:-1])
+            # equal sign => assignment expresion
+            if Syntax.is_assignment(parsed_line):
+                key, value = parsed_line.key_value()
+
+                dotty_key = cls.key(key)
+                final_key = f"{table}{cls._SEPARATOR}{dotty_key}" if table else dotty_key
+
+                data[final_key] = cls.value(value, parsed_line)
+
+            # no equal sign => table assignment, ie: [table]
+            else:
+                # remove "[" and "]", handle quotes/dots
+                table = cls.key(parsed_line.line[1:-1]).replace(".", cls._SEPARATOR)
+
+        return data
+
+
+class Syntax:
+    """Tiny helpers for syntax."""
+
+    @staticmethod
+    def check(__parsed: ParsedLine) -> Optional[str]:
+        """Run some checks."""
+
+        #######################
+        # Table or assignment #
+        #######################
+        is_assignment = Syntax.is_assignment(__parsed)
+        is_table_setter = Syntax.is_in_brackets(__parsed.line)
+
+        if not is_assignment and not is_table_setter:
+            return "Line has to contain either an assignment or table setter"
+
+        if is_assignment and is_table_setter:
+            return "Line cant be an assignment and table setter at the same time"
+
+        ##############
+        # Assignment #
+        ##############
+        if is_assignment and not len(__parsed.line) > (__parsed.tokens[Tokens.EQUAL_SIGN][0] + 1):
+            return "Invalid assignment, nothing after equal sign"
+
+        # If we got here, everything was correct
+        # Empty string => No exception raised
+        return ""
+
+
+    @staticmethod
+    def is_quoted(__val: str) -> bool:
+        """Check if a string is quoted."""
+        return __val[0] == __val[-1] and __val[0] in Tokens.QUOTES
+
+    @staticmethod
+    def is_assignment(__parsed: ParsedLine) -> bool:
+        """Whether this line contains an assignment."""
+        return bool(__parsed.tokens[Tokens.EQUAL_SIGN])
+
+    @staticmethod
+    def is_in_brackets(__val: str) -> bool:
+        return (
+            __val[0] == Tokens.OPENING_BRACKET
+            and __val[-1] == Tokens.CLOSING_BRACKET
+        )
 
 
 ##############
@@ -347,7 +438,7 @@ class Parser:
 ##############
 def loads(__str: str) -> Dotty:
     """Parse TOML from a string."""
-    return Parser(__str).data
+    return Parser.toml(__str)
 
 
 def load(__file: "File") -> Dotty:
@@ -363,20 +454,24 @@ def dumps(__data: Dotty | dict) -> str:
 
     # enclose on a dict, to easily find the "tables" on it
     if isinstance(__data, dict):
-        __data = Dotty(__data, fill_tables=True)
-
-    def _order(x):
-        """Little helper to sort the tables."""
-        return (
-            # len cant be -1 on a string, this ensures root elements being first
-            -1
-            if x == __data._BASE
-            else len(x)
-        )
+        __data = Dotty(__data, fill_tables=True, separator=Parser._SEPARATOR)
 
     out = StringIO()
-    for table_name in sorted(__data.tables, key=_order):
-        table = __data[table_name]
+    for table_parts in __data.tables:
+        if table_parts != __data._BASE:
+            # convert the parts into a full name
+            table_name = ""
+            for part in table_parts:
+                # an empty string is a valid value...
+                if part == "":
+                    part = '""'
+
+                table_name += part + __data._separator
+            table_name = table_name.rstrip(__data._separator)
+
+            table = __data[table_name]
+        else:
+            table = __data[__data._BASE]
 
         # special case for tables without direct childs (ie: only nested ones)
         # skip them
@@ -384,7 +479,7 @@ def dumps(__data: Dotty | dict) -> str:
             continue
 
         # special case for items at root of the dict
-        if table_name != __data._BASE:
+        if table_parts != __data._BASE:
             out.write(f"[{table_name}]\n")
 
         for key, value in table.items():
@@ -393,9 +488,12 @@ def dumps(__data: Dotty | dict) -> str:
                 continue
 
             # enclose string in quotes to prevent casting it when reading
-            # actually, TOML enforces the use of quotes, while this lib does not
             if isinstance(value, str):
                 value = f'"{value}"'
+
+            # if key contains a dot, quote it to maintain the data format
+            if "." in key or key == "":
+                key = f'"{key}"'
 
             out.write(f"{key}={value}\n")